swap tag-range and list-range
