+++ src/list_range.rs
@@ -1,16 +1,10 @@
 use crate::internal::{Arena, Label, PriorityRef};
 pub use crate::MaintainedOrd;
-use order_maintenance_macros::generate_capacities;
 use std::cmp::Ordering;
-
-generate_capacities! {
-    /// Capacities for 17 thresholds in the range `(1.1..=1.9)` (inclusive) with 64-bit tags.
-    const CAPACITIES: [[1.1..=1.9; 64]; 17];
-}
 
 /// A totally-ordered priority.
 ///
-/// These priorities implement Bender et al. (2002)'s solution to the order maintenance problem,
+/// These priorities implement Dietz & Sleator (1987)'s solution to the order maintenance problem,
 /// which require a data structure `T` that supports insertion and comparison operations such that
 /// insertion constructs an element of the next greatest priority:
 ///
@@ -43,124 +37,95 @@ generate_capacities! {
 /// assert!(p1 < p3);
 /// assert!(p2 < p3);
 /// ```
+///
+/// ## Memory management
+///
+/// Under the hood, these priorities are actually references to nodes of a circular linked list,
+/// allocated from an arena. Those nodes are reference-counted, which allows these priorities to be
+/// cloned. The node's reference count is decremented when this priority is dropped, and if it
+/// reaches zero, the node is deallocated.
+///
+/// Priorities from different arenas cannot be compared with one another.
+///
+/// ## Algorithm
+///
+/// This implementation uses Dietz & Sleator (1987)'s algorithm, also called list-range relabeling
+/// (as opposed to Bender et al.'s list-range relabeling algorithm).
+///
+/// While Dietz & Sleator also propose a data structure that supports constant-time insertion, that
+/// data structure is so overwhelmingly complex that the overhead of maintaining such a data
+/// structure will overwhelm any theoretical efficiency for any reasonable number of priorities.
+///
+/// More recently, Bender et al. proposed an alternative solution, using a list-range relabling
+/// approach. That approach is likely more efficient on real hardware, since it favors bit-wise
+/// operations over multiplication and division. For now, this crate uses the possibly slower
+/// list-range relabeling approach, because it was ported from a scripting language that is better
+/// suited toward floating operations. It remains to be seen which implementation is better under
+/// which circumstances.
+///
+/// ## References
+///
+/// -   Paul F. Dietz and Daniel D. Sleator. _Two Algorithms for Maintaining Order in a List._ 1987.
+///
+/// -   Michael A. Bender, Richard Cole, Erik D. Demaine, Martin Farach-Colton, and Jack Zito.
+///     _Two simplified algorithms for maintaining order in a list._ 2002.
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct Priority(PriorityRef);
 
 impl Priority {
     fn relative(&self) -> Label {
-        self.0.label()
+        self.0.label() - self.0.base_label()
     }
 
-    /// Find the correct list of capacities depending onnumber of priorities already inserted.
-    fn threshold_index(&self, total: usize) -> usize {
-        // find the correct list of capacities depending onnumber of priorities already inserted
-        for (i, _) in CAPACITIES.iter().enumerate().rev() {
-            let last = *unsafe { CAPACITIES[i].last().unwrap_unchecked() };
-            if total + 1 < last {
-                return i;
-            }
+    /// Search for how many nodes we need to relabel, and its weight
+    fn check_label_range(&self, arena: &mut Arena) -> (usize, Label) {
+        let this = self.0.this().as_ref(arena);
+        let mut count = 1;
+        let mut prio = this.next().as_ref(arena);
+
+        let mut weight = prio.label() - this.label();
+        while weight != 0 && weight <= count * count {
+            prio = prio.next().as_ref(arena);
+            count += 1;
+            weight = prio.label() - this.label();
         }
-
-        panic!("Too many priorities were inserted: {total}");
+        (count, weight)
     }
 
-    /// Perform relabeling in the arena.
-    fn do_relabel(&self, arena: &mut Arena) {
+    fn redistribute_labels(&self, arena: &mut Arena, count: usize, weight: Label) {
         let this = self.0.this().as_ref(arena);
 
-        let t_index = self.threshold_index(arena.total());
-
-        let mut i = 0;
-        let mut range_size = 1;
-        let mut range_count = 1;
-        let mut internal_node_tag = this.label();
-
-        // the subrange is [min_lab, max_lab] (inclusive)
-        let mut min_lab = internal_node_tag;
-        let mut max_lab = internal_node_tag;
-
-        let mut begin = this;
-        let mut end = this;
-
-        // The density threshold is 1/T^i
-        // So we want to find the smallest subrange so that count/2^i <= 1/T^i
-        // or count <= (2/T)^i = CAPA[t_index][i]
-
-        while range_size < usize::MAX {
-            loop {
-                let new_begin = begin.prev().as_ref(arena);
-                if new_begin.label() < min_lab || new_begin.label() >= begin.label() {
-                    break;
-                }
-                range_count += 1;
-                begin = new_begin;
-            }
-            loop {
-                let new_end = end.next().as_ref(arena);
-                if new_end.label() > max_lab || new_end.label() <= end.label() {
-                    break;
-                }
-                range_count += 1;
-                end = new_end;
-            }
-
-            if range_count < CAPACITIES[t_index][i] {
-                // Range found, relabel
-                let gap = range_size / range_count;
-                let mut rem = range_size % range_count; // note: the reminder is spread out
-                let mut new_label = min_lab;
-
-                while begin.label() != end.label() {
-                    begin.set_label(new_label);
-                    begin = begin.next().as_ref(arena);
-                    new_label += gap;
-                    if rem > 0 {
-                        new_label += 1;
-                        rem -= 1;
-                    }
-                }
-                end.set_label(new_label); // the end is part of the range
-
-                break;
+        // Now, adjust labels of those nodes
+        let mut prio = this.next().as_ref(arena);
+        for k in 1..count {
+            // if weight == 0, then it should actually encode usize::MAX + 1.
+            let weight_k = if weight == 0 {
+                // Since we can't actually represent usize::MAX + 1, we just multiply it by
+                // ((usize::MAX + 1) / 2) AKA (1 << (usize::BITS / 2)), and then multiply by 2.
+                Label::new((k * (1 << (Label::BITS / 2))) * 2)
             } else {
-                if range_size == usize::MAX {
-                    panic!("Too many priorities were inserted, the root is overflowing!");
-                }
-                i += 1;
-                range_size *= 2;
-                internal_node_tag >>= 1;
-                min_lab = internal_node_tag << i;
-                max_lab = (internal_node_tag + 1) << i;
-            }
+                weight * k
+            };
+            prio.set_label(weight_k / count + this.label());
+
+            prio = prio.next().as_ref(arena);
         }
     }
 
     /// Perform relabeling in the arena if necessary.
     fn relabel(&self, arena: &mut Arena) {
-        let this = self.0.this().as_ref(arena);
-        let next = this.next().as_ref(arena);
-        let next_lab = if next.label() <= this.label() {
-            Label::MAX
-        } else {
-            next.label()
-        };
-
-        if this.label() + 1 == next_lab {
-            self.do_relabel(arena)
+        // Search for how many nodes we need to relabel, and its weight
+        let (count, weight) = self.check_label_range(arena);
+        if count > 1 {
+            self.redistribute_labels(arena, count, weight);
         }
     }
 
     /// Compute the next label for inserting after `self`.
     fn next_label(&self, arena: &Arena) -> Label {
         let this = self.0.this().as_ref(arena);
-        let next = this.next().as_ref(arena);
-        let next_lab = if next.label() <= this.label() {
-            Label::MAX
-        } else {
-            next.label()
-        };
-
-        (this.label() & next_lab) + ((this.label() ^ next_lab) >> 1)
+        // Compute new priority, which is half-way between this priority and the next
+        this.label() + (this.next().as_ref(arena).label() - this.label()) / 2
     }
 }
 
@@ -178,11 +143,11 @@ impl PartialOrd for Priority {
 
 impl MaintainedOrd for Priority {
     fn new() -> Self {
-        let arena = Arena::new();
-        // Base is not a specially designated priority in this implementation, so we
-        // can use it as the first priority.
-        let this = arena.base();
-        Priority(PriorityRef::new(arena, this))
+        let mut arena = Arena::new();
+
+        // For list-range, the base is a special priority, so we need to use another one.
+        let this = arena.insert_after(Label::MAX / 2, arena.base());
+        Self(PriorityRef::new(arena, this))
     }
 
     fn insert(&self) -> Self {
